backend/requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
python-dotenv==1.0.0
python-multipart==0.0.6
requests==2.31.0
twilio==8.5.0
google-generativeai==0.3.1
pillow==10.0.1
pydantic==2.4.2
python-jose==3.3.0
pydantic-settings==2.10.1
pytesseract==0.3.10
opencv-python==4.8.0.74
numpy==1.24.3
symspellpy==6.7.7
nltk==3.8.1
spacy==3.7.2
sentence-transformers==2.2.2
faiss-cpu==1.7.4
transformers==4.36.2
torch==2.1.2
reportlab==4.0.8
polyglot==16.7.4
pyicu==2.12
pycld2==0.41
morfessor==2.0.6
psutil==5.9.6